requirements.txt
cryptography==2.8
openssl-python==0.1.1
pika==1.1.0
Pillow==7.0.0
pycryptodome==3.9.4
pylint==2.4.2
pyOpenSSL==19.1.0
python-ldap==3.2.0
